Add GitHub Actions for testing and releasing
Added two new GitHub workflows. The first, 'Test', is set up to run tests on push to the master branch. This ensures that the project is always in a working state. The second, 'Publish package to the Maven Central Repository', is triggered on tag push and is responsible for releasing the project to Maven Central.